perubahan 14/01

+++ app/Http/Controllers/Admin/MasterData/MasterJenisHewanController.php
@@ -1,41 +1,37 @@
 <?php
 
-namespace App\Http\Controllers\Admin\WebContent;
+namespace App\Http\Controllers\Admin\MasterData;
 
 use Exception;
 use App\Helper;
-use App\Models\Faq;
-use Illuminate\Support\Str;
 use Illuminate\Http\Request;
+use App\Models\MasterJenisHewan;
 use Illuminate\Support\Facades\DB;
 use App\Http\Controllers\Controller;
 use Yajra\DataTables\Facades\DataTables;
 use Illuminate\Support\Facades\Validator;
 
-class FaqController extends Controller
+class MasterJenisHewanController extends Controller
 {
-    public $title = 'FAQ';
+    public $title = 'Master Jenis Hewan';
 
     public function index()
     {
-        return view('admin.web-content.faq',[
-            'title' => $this->title
+        return view('admin.master-data.master-jenis-hewan',[
+            'title' => $this->title,
         ]);
     }
 
     public function data(Request $request)
     {
         if ($request->ajax()) {
-            $data = Faq::latest()->get();
+            $data = MasterJenisHewan::latest()->get();
             return DataTables::of($data)
                 ->addIndexColumn()
                 ->addColumn('action', function($row){
-                    $action = '<button type="button" data-id="'.$row->id.'" class="btn btn-primary btn-sm my-1 edit">Edit</button>
-                                <button type="button" data-id="'.$row->id.'" class="btn btn-danger btn-sm delete">Hapus</button>';
+                    $action = '<button data-id="'.$row->id.'" class="btn btn-primary btn-sm my-1 edit">Edit</button>
+                                <button data-id="'.$row->id.'" class="btn btn-danger btn-sm delete">Hapus</button>';
                     return $action;
-                })
-                ->editColumn('jawaban', function($row){
-                    return Str::limit($row->jawaban,30);
                 })
                 ->rawColumns(['action'])
                 ->make(true);
@@ -45,8 +41,8 @@ class FaqController extends Controller
     public function store(Request $request)
     {
         $validator = Validator::make($request->all(), [
-            'pertanyaan' => 'required|unique:faq',
-            'jawaban' => 'required',
+            'nama' => 'required|unique:master_jenis_hewan',
+            'deskripsi' => 'nullable',
         ]);
         if ($validator->fails()) {
             return response()->json([
@@ -55,11 +51,11 @@ class FaqController extends Controller
         } else {
             try {
                 DB::beginTransaction();
-                Faq::create([
-                    'pertanyaan' => $request->pertanyaan,
-                    'jawaban' => $request->jawaban
+                $jenis = MasterJenisHewan::create([
+                    'nama' => $request->nama,
+                    'deskripsi' => $request->deskripsi,
                 ]);
-                Helper::createUserLog("Berhasil menambahkan FAQ", auth()->user()->id, $this->title);
+                Helper::createUserLog("Berhasil menambahkan jenis hewan ".$jenis->nama, auth()->user()->id, $this->title);
                 DB::commit();
                 return response()->json([
                     'status'=>200,
@@ -67,7 +63,7 @@ class FaqController extends Controller
                 ]);
             } catch (Exception $e) {
                 DB::rollBack();
-                Helper::createUserLog("Gagal menambah FAQ", auth()->user()->id, $this->title);
+                Helper::createUserLog("Gagal menambah jenis hewan", auth()->user()->id, $this->title);
                 return response()->json([
                     'message'=>$e->getMessage()
                 ],422);
@@ -77,7 +73,7 @@ class FaqController extends Controller
 
     public function edit($id)
     {
-        $data = Faq::find($id);
+        $data = MasterJenisHewan::find($id);
         if ($data) {
             return response()->json([
                 'status' => 200,
@@ -90,11 +86,11 @@ class FaqController extends Controller
         }
     }
 
-    public function update(Request $request)
+    public function update(Request $request, $id)
     {
         $validator = Validator::make($request->all(), [
-            'pertanyaan' => 'required',
-            'jawaban' => 'required',
+            'nama' => 'required',
+            'deskripsi' => 'nullable',
         ]);
         if ($validator->fails()) {
             return response()->json([
@@ -103,30 +99,30 @@ class FaqController extends Controller
         } else {
             try {
                 DB::beginTransaction();
-                $data = Faq::find($request->id);
-                if ($data->pertanyaan==$request->pertanyaan) {
+                $data = MasterJenisHewan::find($id);
+                if ($data->nama==$request->nama) {
                     $data->update([
-                        'pertanyaan' => $request->pertanyaan,
-                        'jawaban' => $request->jawaban
+                        'nama' => $request->nama,
+                        'deskripsi' => $request->deskripsi,
                     ]);
-                    Helper::createUserLog("Berhasil mengubah FAQ", auth()->user()->id, $this->title);
+                    Helper::createUserLog("Berhasil mengubah jenis hewan", auth()->user()->id, $this->title);
                     DB::commit();
                     return response()->json([
                         'status'=>200,
                         'message'=>'Berhasil mengubah data'
                     ]);
                 } else {
-                    $cek = Faq::where('pertanyaan',$request->pertanyaan)->whereNot('id',$request->id)->first();
+                    $cek = MasterJenisHewan::where('nama',$request->nama)->whereNot('id',$id)->first();
                     if ($cek) {
                         return response()->json([
-                            'message'=>'Pertanyaan sudah digunakan!'
+                            'message'=>'Nama sudah digunakan!'
                         ],404);
                     } else {
                         $data->update([
-                            'pertanyaan' => $request->pertanyaan,
-                            'jawaban' => $request->jawaban
+                            'nama' => $request->nama,
+                            'deskripsi' => $request->deskripsi,
                         ]);
-                        Helper::createUserLog("Berhasil mengubah FAQ", auth()->user()->id, $this->title);
+                        Helper::createUserLog("Berhasil mengubah jenis hewan ", auth()->user()->id, $this->title);
                         DB::commit();
                         return response()->json([
                             'status'=>200,
@@ -135,7 +131,7 @@ class FaqController extends Controller
                     }
                 }
             } catch (Exception $e) {
-                Helper::createUserLog("Gagal mengubah FAQ", auth()->user()->id, $this->title);
+                Helper::createUserLog("Gagal mengubah data jenis hewan", auth()->user()->id, $this->title);
                 return response()->json([
                     'message'=>$e->getMessage()
                 ],422);
@@ -145,18 +141,18 @@ class FaqController extends Controller
 
     public function destroy($id)
     {
-        $data = Faq::find($id);
-        try {
-            Helper::createUserLog("Berhasil menghapus FAQ", auth()->user()->id, $this->title);
+        $data = MasterJenisHewan::find($id);
+        if ($data) {
+            Helper::createUserLog("Berhasil menghapus jenis hewan ".$data->nama, auth()->user()->id, $this->title);
             $data->delete();
             return response()->json([
                 'status'=>200,
                 'message'=>'Berhasil menghapus data'
             ]);
-        } catch (Exception $e) {
-            Helper::createUserLog("Gagal menghapus FAQ", auth()->user()->id, $this->title);
+        } else {
+            Helper::createUserLog("Gagal menghapus data jenis hewan", auth()->user()->id, $this->title);
             return response()->json([
-                'message'=>$e->getMessage()
+                'message'=>'Gagal menghapus'
             ],422);
         }
     }
